test: Consolidate COVERAGE tests into tests/unit/ and drop root-level duplicate
tests/test_coverage.py duplicated TestGIQLCoverage (also present in
tests/unit/test_expressions.py) and TestCoverageTransformer (also
present in tests/unit/test_transformer.py) with stronger content in
the root file — more tests, proper AAA phase comments, proper
Given-When-Then docstrings, property tests, and end-to-end DuckDB
coverage. The root file also violated the test guide's mirror-the-
source-module convention because there is no src/giql/coverage.py
module.

Replace the weaker TestGIQLCoverage in tests/unit/test_expressions.py
with the 10-test version from the root file (up from 7). Replace the
weaker TestCoverageTransformer in tests/unit/test_transformer.py
with the 37-test version (up from 20). Carry over the Hypothesis,
duckdb, Table, and transpile imports needed by the merged content,
plus the VALID_STATS module constant. The to_df fixture already
lives in tests/conftest.py so it is auto-discovered by tests/unit/
without a new conftest.

Delete tests/test_coverage.py.

Net: +47 coverage tests in unit/ - 27 weaker duplicates - 1222 lines
of the root file = 89 tests in the two touched unit files after the
merge (up from ~57).
